add plugins support
